requirements.txt
# Core dependencies
openai>=1.0.0
numpy>=1.24.0
cma>=3.3.0
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Optional for visualization/analysis
pandas>=2.0.0
tabulate>=0.9.0